// src/types.ts
export interface IDXBasicProfile {
  name?: string
  image?: string
  description?: string
  emoji?: string
  background?: string
  birthDate?: string
  url?: string
  gender?: string
  homeLocation?: string
  residenceCountry?: string
  nationalities?: Array<string>
  affiliations?: Array<string>
}